Apulu hitless, dol working (#26509)

+++ nic/apollo/tools/apulu/upgrade/start-upgmgr-mock.sh
@@ -5,19 +5,20 @@ CUR_DIR=$( readlink -f $( dirname $0 ))
 
 source  ${CUR_DIR}/../../../tools/setup_env_mock.sh $PIPELINE
 
-export OPERD_REGIONS=$CONFIG_PATH/operd-regions.json
-
-echo "Starting pdsupgmgr: `date +%x_%H:%M:%S:%N`"
-
-rm -f $CONFIG_PATH/operd-regions.json
-ln -s $CONFIG_PATH/$PIPELINE/operd-regions.json $CONFIG_PATH/operd-regions.json
+if [ ! -e $CONFIG_PATH/operd-regions.json ];then
+    ln -s $CONFIG_PATH/$PIPELINE/operd-regions.json $CONFIG_PATH/operd-regions.json
+    export OPERD_REGIONS=$CONFIG_PATH/operd-regions.json
+fi
 
 function finish {
     echo "pdsupgmgr exit"
-    operdctl dump upgradelog > upgrade.log
-    rm -f $CONFIG_PATH/operd-regions.json
+    operdctl dump upgradelog > upgrademgr.log
+    if [ ! -z "$OPERD_REGIONS" ];then
+        rm -f $CONFIG_PATH/operd-regions.json
+    fi
 }
 trap finish EXIT
 
+echo "Starting pdsupgmgr: `date +%x_%H:%M:%S:%N`"
 CMD="$BUILD_DIR/bin/pdsupgmgr $* 2>&1 &"
 exec $CMD
